Modify model/UserModel.py
