+++ app.py
@@ -39,16 +39,7 @@ app.state.db = myDB
 
 
 
-# 建立 user 資料 model
-class User(BaseModel):
-    id: int | None
-    name: str | None
-    email: EmailStr | None
 
-
-# 建立 userInfo 資料 model
-class UserInfo(BaseModel):
-    data: User | None = None
 
 
 # 建立 signUpInfo 資料 model
